edit : StompLogHandler -> WebSocket 내 인증 & 예외 처리 삭제, 로그 기록을 위한 핸들러로 변경
